requirements.txt
alembic==1.11.2
gunicorn==20.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-Session==0.8.0
Flask-SQLAlchemy==3.0.2
Flask-SSLify==0.1.5
flask-swagger-ui==4.11.1
marshmallow==3.20.2
psycopg2-binary==2.9.6
pytest==7.1.3
python-dotenv==0.21.1
requests==2.11.1
SQLAlchemy==1.4.45
Werkzeug==2.2.2
